Added the order process of the app

+++ src/pages/SignupPage.js
@@ -1,30 +1,60 @@
-import {React,useEffect, useContext, useState,Link, useLocation, useNavigate,Form,Button,Container,axios,store,toast,getError,Title,USER_SIGNIN } from '../Imports';
+import { toast } from 'react-toastify';
+import {
+  axios,
+  useLocation,
+  useNavigate,
+  Container,
+  useContext,
+  useEffect,
+  useState,
+  store,
+  getError,
+  Title,
+  USER_SIGNIN,
+  Form,
+  Link,
+  Button,
+} from '../Imports';
 
-export default function SignInPage() {
+function SignupPage() {
   const navigate = useNavigate();
   const { search } = useLocation();
   const redirectInUrl = new URLSearchParams(search).get('redirect');
   const redirect = redirectInUrl ? redirectInUrl : '/';
 
+  const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, SetConfirmPassword] = useState('');
 
   const { state, dispatch: ctxDispatch } = useContext(store);
   const { userInfo } = state;
+
   const submitHandler = async (e) => {
     e.preventDefault();
+
+    if (password !== confirmPassword) {
+      toast.error('Passwords must match!');
+      return;
+    }
+
     try {
-      const { data } = await axios.post('/api/v1/users/signin', {
+      console.log(name, email, password);
+      const { data } = await axios.post('/api/v1/users/signup', {
+        name,
         email,
         password,
       });
+
       ctxDispatch({ type: USER_SIGNIN, payload: data });
+
       localStorage.setItem('userInfo', JSON.stringify(data));
       navigate(redirect || '/');
     } catch (err) {
-        toast.error(getError(err));
+      toast.error(getError(err));
     }
   };
+
   useEffect(() => {
     if (userInfo) {
       navigate(redirect);
@@ -33,9 +63,17 @@ export default function SignInPage() {
 
   return (
     <Container className="small-container">
-      <Title title="Sign In" />
-      <h1 className="my-3">Sign In</h1>
-      <Form onSubmit={(e) => submitHandler(e)}>
+      <Title title="Sign-up" />
+      <h1 className="my-3">Sign Up</h1>
+      <Form onSubmit={submitHandler}>
+        <Form.Group className="mb-3" controlId="name">
+          <Form.Label>Name</Form.Label>
+          <Form.Control
+            type="text"
+            required
+            onChange={(e) => setName(e.target.value)}
+          />
+        </Form.Group>
         <Form.Group className="mb-3" controlId="email">
           <Form.Label>Email</Form.Label>
           <Form.Control
@@ -52,17 +90,23 @@ export default function SignInPage() {
             onChange={(e) => setPassword(e.target.value)}
           />
         </Form.Group>
+        <Form.Group className="mb-3" controlId="confirmPassword">
+          <Form.Label>Confirm Password</Form.Label>
+          <Form.Control
+            type="password"
+            required
+            onChange={(e) => SetConfirmPassword(e.target.value)}
+          />
+        </Form.Group>
         <div className="mb-3">
-          <Button type="submit">Sign In</Button>
+          <Button type="submit">Sign Up</Button>
         </div>
         <div className="mb-3">
-          New customer?{' '}
-          <Link to={`/signup?redirect=${redirect}`}>Create your account</Link>
-        </div>
-        <div className="mb-3">
-          Forget Password? <Link to={'/forgot-password'}></Link>
+          Already have an account?{' '}
+          <Link to={`/signin?redirect=${redirect}`}>Sign in</Link>
         </div>
       </Form>
     </Container>
   );
 }
+export default SignupPage;
